Fix a panic, optimized cpu usage a bit (mainly in tests)
The panic was caused by the fact that we were overwriting unit's Terrain field with its own sprite, because we were assuming it moved to a new spot, which is not necessarily the case.
Unit struct is big enough that looping _, unit := range units is way
slower than i := range units if we're mostly touching only a few of the
unit's fields.