Fix for bug in Crusade/Campaign data file
The Utah Beach is marked to be occupied by Germans, but there's an allied unit in the city.
It messes up with game's AI.
Add a check that at the start of the game, checks if there are cities that should be immediately captured.
It makes the game behaviour different than in the original game, but it seems an obvious bug that needs to be fixed.